Financial Model Update

+++ app/containers/Pages/ActionType/ActionTypeBlock.js
@@ -4,24 +4,41 @@ import IconButton from '@material-ui/core/IconButton';
 import DetailsIcon from '@material-ui/icons/Details';
 import DeleteIcon from '@material-ui/icons/Delete';
 import {
-  Button, Dialog, DialogActions, DialogContent, DialogTitle, Grid, TextField
+  Button,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogTitle,
+  FormControl,
+  Grid,
+  InputLabel, MenuItem,
+  Select,
+  TextField,
+  Typography
 } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import { connect } from 'react-redux';
 import CustomToolbar from '../../../components/CustomToolbar/CustomToolbar';
 import ActionTypeService from '../../Services/ActionTypeService';
 import { ThemeContext } from '../../App/ThemeWrapper';
+import CommercialActionService from '../../Services/CommercialActionService';
 
 const useStyles = makeStyles();
 
 class ActionTypeBlock extends React.Component {
   constructor(props) {
     super(props);
+    // eslint-disable-next-line react/destructuring-assignment,react/prop-types
     const thelogedUser = JSON.parse(this.props.logedUser);
     this.state = {
       actionTypeId: '',
+      id: '',
+      newActionTypeId: '',
       datas: [],
+      commercialActions: [],
+      newTab: [],
       openPopUp: false,
+      openWarning: false,
       typeName: '',
       description: '',
       columns: [
@@ -70,6 +87,9 @@ class ActionTypeBlock extends React.Component {
     ActionTypeService.getActionType().then(result => {
       this.setState({ datas: result.data });
     });
+    CommercialActionService.getCommercialAction().then(result => {
+      this.setState({ commercialActions: result.data.payload });
+    });
     const {
       // eslint-disable-next-line react/prop-types
       changeTheme
@@ -94,18 +114,29 @@ class ActionTypeBlock extends React.Component {
     }
 
     handleDelete = (tableMeta) => {
+      let test = false;
+      let newTab = [];
       const index = tableMeta.tableState.page * tableMeta.tableState.rowsPerPage
             + tableMeta.rowIndex;
       // eslint-disable-next-line react/destructuring-assignment,react/no-access-state-in-setstate
       const id = this.state.datas[index].actionTypeId;
-      console.log(id);
-      ActionTypeService.deleteActionType(id).then(result => {
-        this.setState({ datas: result.data });
+      // eslint-disable-next-line array-callback-return,react/destructuring-assignment
+      this.state.commercialActions.map(row => {
+        if ((row.commercialActionType._id) === (id)) test = true;
       });
+      if (test) {
+        // eslint-disable-next-line react/destructuring-assignment
+        newTab = this.state.datas.filter(row => row.actionTypeId !== id);
+        this.setState({ openWarning: true, newTab, id });
+      } else {
+        ActionTypeService.deleteActionType(id).then(result => {
+          this.setState({ datas: result.data });
+        });
+      }
     };
 
     handleClose = () => {
-      this.setState({ openPopUp: false });
+      this.setState({ openPopUp: false, openWarning: false });
     };
 
     handleSave = () => {
@@ -120,17 +151,34 @@ class ActionTypeBlock extends React.Component {
       });
     };
 
+    handleReplaceAction = () => {
+      const {
+        newActionTypeId, id
+      } = this.state;
+      // eslint-disable-next-line react/destructuring-assignment,array-callback-return
+      this.state.commercialActions.map(row => {
+        if ((row.commercialActionType._id) === (id)) {
+          row.commercialActionType._id = newActionTypeId;
+          CommercialActionService.updateCommercialAction(row).then(result => {
+            console.log(result);
+          });
+        }
+      });
+      ActionTypeService.deleteActionType(id).then(result => {
+        this.setState({ datas: result.data });
+      });
+    }
+
     handleChange = (ev) => {
       this.setState({ [ev.target.name]: ev.target.value });
     };
 
     render() {
       const {
-        columns, openPopUp, datas, typeName, description
+        columns, openPopUp, datas, typeName, description, openWarning, newActionTypeId, newTab
       } = this.state;
-      const {
-        logedUser
-      } = this.props;
+      // eslint-disable-next-line react/prop-types
+      const { logedUser } = this.props;
       const thelogedUser = JSON.parse(logedUser);
       let exportButton = false;
       if (thelogedUser.userRoles[0].actionsNames.financialModule_typeOfCurrency_export) {
@@ -224,6 +272,67 @@ class ActionTypeBlock extends React.Component {
                             save
                 </Button>
               ) : null}
+            </DialogActions>
+          </Dialog>
+          <Dialog
+            open={openWarning}
+            keepMounted
+            scroll="paper"
+            onClose={this.handleClose}
+            aria-labelledby="alert-dialog-slide-title"
+            aria-describedby="alert-dialog-slide-description"
+            fullWidth=""
+            maxWidth=""
+          >
+            <DialogTitle id="alert-dialog-slide-title"> Operation Denied </DialogTitle>
+            <DialogContent dividers>
+              <Typography
+                style={{
+                  color: '#000',
+                  fontFamily: 'sans-serif , Arial',
+                  fontSize: '16px',
+                  fontWeight: 'bold',
+                  opacity: 0.4,
+                  marginRight: 20,
+                  width: '100%'
+                }}
+              >
+                This Action type is used in other commercial actions. If you want to force delete,
+                select other action type to be replaced with.
+              </Typography>
+              <Grid
+                container
+                spacing={2}
+                alignItems="flex-start"
+                direction="row"
+                justify="center"
+              >
+                <Grid item xs={4}>
+                  <FormControl fullWidth required>
+                    <InputLabel>Commercial Action Type</InputLabel>
+                    <Select
+                      name="newActionTypeId"
+                      value={newActionTypeId}
+                      onChange={this.handleChange}
+                    >
+                      {newTab.map((clt) => (
+                        <MenuItem key={clt.actionTypeId} value={clt.actionTypeId}>
+                          {clt.typeName}
+                        </MenuItem>
+                      ))
+                      }
+                    </Select>
+                  </FormControl>
+                </Grid>
+              </Grid>
+            </DialogContent>
+            <DialogActions>
+              <Button color="secondary" onClick={this.handleClose}>
+                Cancel
+              </Button>
+              <Button variant="contained" color="primary" onClick={this.handleReplaceAction}>
+                Save
+              </Button>
             </DialogActions>
           </Dialog>
         </div>
